Fix sbin/ build with a 64-bit ino_t.
Original code by:	Gleb Kurtsou


git-svn-id: svn+ssh://svn.freebsd.org/base/head@241013 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sbin/newfs/mkfs.c
@@ -1003,7 +1003,8 @@ iput(union dinode *ip, ino_t ino)
 	sblock.fs_cstotal.cs_nifree--;
 	fscs[0].cs_nifree--;
 	if (ino >= (unsigned long)sblock.fs_ipg * sblock.fs_ncg) {
-		printf("fsinit: inode value out of range (%d).\n", ino);
+		printf("fsinit: inode value out of range (%ju).\n",
+		    (uintmax_t)ino);
 		exit(32);
 	}
 	d = fsbtodb(&sblock, ino_to_fsba(&sblock, ino));
